refactor: use-data-loader hook file structure
