+++ src/pages/Homepage/index.test.tsx
@@ -3,10 +3,10 @@ import '@src/__mocks__/loader-mock';
 import '@src/__mocks__/books-service-context-mock';
 import { screen, render, fireEvent, waitFor } from '@testing-library/react';
 import Homepage from '.';
-import { Loading } from '@src/hooks/use-data-loader';
+import { Loading } from '@src/hooks/use-data-loader/types';
 import * as dataLoader from '@src/hooks/use-data-loader';
 
-const dataLoaderMock = dataLoader as { Loading: object; useDataLoader: () => void };
+const dataLoaderMock = dataLoader as { useDataLoader: () => void };
 
 const loadDataMock = jest.fn(async (fetchData: () => Promise<unknown>, setData: () => void) => {
   await fetchData();
@@ -15,7 +15,6 @@ const loadDataMock = jest.fn(async (fetchData: () => Promise<unknown>, setData: 
 
 jest.mock('@src/hooks/use-data-loader', () => ({
   __esModule: true,
-  Loading: { IDLE: 0, PENDING: 1, SUCCESS: 2, ERROR: 3 },
   useDataLoader: jest.fn(),
 }));
